Updating for HW5
This is the final submission of Joel Wilhelm and Dallas Barnett’s AI
final project aka homework 5. We fixed the issues that our professor,
Dr. Nancy Green, found in homework 4 and fixed the corresponding code.
In addition to that, we combined the forward chaining and BN together
to form the master program.